Shard task-shape contracts by language
