+++ code/packages/typescript/human-language-data/tests/task-shapes.test.ts
@@ -3,357 +3,33 @@ import { loadLanguageRegistry, loadTaskShapeInventory, listTaskShapeInventories 
 import { buildTaskShapeBacklog, parseTaskShapeInventory } from "../src/task-shapes.js";
 
 describe("four-skill task-shape inventories (HL18)", () => {
-  it("loads the project-defined Gujarati pre-A1 floor with four separate 100-point papers", () => {
-    const inventory = loadTaskShapeInventory("gujarati", "pre-A1");
-    expect(inventory.target).toEqual({
-      name: "Coding Adventures Gujarati pre-A1 Assessment — project-defined equivalent",
-      basis: "project-defined",
-    });
-    expect(inventory.sections.map((section) => section.skill)).toEqual([
-      "reading",
-      "listening",
-      "writing",
-      "speaking",
-    ]);
-    expect(inventory.administration).toMatchObject({
-      writtenMinutes: 32,
-      speakingMinutes: 8,
-      speakingPreparationMinutes: 0,
-    });
-    expect(inventory.sections.map((section) =>
-      section.parts.reduce((sum, part) => sum + (part.scoring.maxRawPoints ?? 0), 0)
-    )).toEqual([100, 100, 100, 100]);
-    expect(inventory.passRule).toMatchObject({ maximumPoints: 400, passPoints: 240 });
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([0.6, 0.6, 0.6, 0.6]);
-    expect(inventory.sections.find((section) => section.skill === "writing")?.parts.map((part) => part.id)).toEqual([
-      "prea1-writing-delayed-recall",
-      "prea1-writing-dictation",
-      "prea1-writing-bounded-production",
-    ]);
-  });
-
-  it("loads the project-defined Marwadi pre-A1 floor as four separate 100-point papers", () => {
-    const inventory = loadTaskShapeInventory("marwadi", "pre-A1");
-    expect(inventory.target).toEqual({
-      name: "Coding Adventures Marwadi pre-A1 Assessment — project-defined equivalent",
-      basis: "project-defined",
-    });
-    expect(inventory.sections.map((section) => section.skill)).toEqual([
-      "reading",
-      "listening",
-      "writing",
-      "speaking",
-    ]);
-    expect(inventory.administration).toMatchObject({
-      writtenMinutes: 32,
-      speakingMinutes: 8,
-      speakingPreparationMinutes: 0,
-    });
-    expect(inventory.sections.map((section) => section.minutes)).toEqual([10, 10, 12, 8]);
-    expect(inventory.sections.map((section) => section.parts.map((part) => part.id))).toEqual([
-      [
-        "prea1-reading-signs-and-words",
-        "prea1-reading-phrases-and-notices",
-        "prea1-reading-personal-details",
-      ],
-      [
-        "prea1-listening-sounds-and-words",
-        "prea1-listening-greeting-responses",
-        "prea1-listening-personal-details",
-      ],
-      [
-        "prea1-writing-delayed-recall",
-        "prea1-writing-dictation",
-        "prea1-writing-bounded-production",
-      ],
-      [
-        "prea1-speaking-greeting-and-identity",
-        "prea1-speaking-familiar-questions",
-        "prea1-speaking-short-request",
-      ],
-    ]);
-    expect(inventory.sections.map((section) => section.parts.map((part) => part.items))).toEqual([
-      [6, 4, 4],
-      [6, 4, 4],
-      [2, 4, 2],
-      [2, 3, 1],
-    ]);
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([0.6, 0.6, 0.6, 0.6]);
-    const paperPoints = inventory.sections.map((section) =>
-      section.parts.reduce((sum, part) => sum + (part.scoring.maxRawPoints ?? 0), 0)
-    );
-    expect(paperPoints).toEqual([100, 100, 100, 100]);
-    expect(paperPoints.reduce((sum, points) => sum + points, 0)).toBe(inventory.passRule.maximumPoints);
-    expect(inventory.passRule).toMatchObject({ maximumPoints: 400, passPoints: 240 });
-  });
-
-  it("loads the project-defined Marwadi A1 destination as an executable four-paper envelope", () => {
-    const inventory = loadTaskShapeInventory("marwadi", "A1");
-    expect(inventory.target).toEqual({
-      name: "Coding Adventures Marwadi A1 Assessment — project-defined equivalent",
-      basis: "project-defined",
-    });
-    expect(inventory.administration).toMatchObject({
-      writtenMinutes: 58,
-      speakingMinutes: 10,
-      speakingPreparationMinutes: 5,
-    });
-    expect(inventory.sections.map((section) => section.minutes)).toEqual([20, 18, 20, 10]);
-    expect(inventory.sections.map((section) => section.parts.map((part) => part.id))).toEqual([
-      ["a1-reading-signs-and-forms", "a1-reading-short-messages", "a1-reading-personal-descriptions"],
-      ["a1-listening-announcements", "a1-listening-short-exchanges", "a1-listening-personal-account"],
-      ["a1-writing-practical-form", "a1-writing-reader-purpose-message"],
-      ["a1-speaking-personal-interview", "a1-speaking-prepared-description", "a1-speaking-transactional-role-play"],
-    ]);
-    expect(inventory.sections.map((section) => section.parts.map((part) => part.items))).toEqual([
-      [7, 7, 6],
-      [6, 6, 5],
-      [1, 1],
-      [5, 1, 1],
-    ]);
-
-    const [reading, listening, writing] = inventory.sections;
-    expect(reading?.parts.reduce((sum, part) => sum + (part.stimulusLength?.minimum ?? 0), 0)).toBe(250);
-    expect(reading?.parts.reduce((sum, part) => sum + (part.stimulusLength?.maximum ?? 0), 0)).toBe(350);
-    expect(reading?.parts.every((part) =>
-      part.promptGenres.every((genre) => genre.includes("no source over 90 words"))
-    )).toBe(true);
-    expect(listening?.parts.every((part) =>
-      part.promptModes.includes("recorded Marwari at 90-110 words per minute") && part.replayCount === 2
-    )).toBe(true);
-    expect(writing?.parts.map((part) => part.responseLength)).toEqual([
-      { unit: "items", minimum: 6, maximum: 8, approximate: false },
-      { unit: "words", minimum: 30, maximum: 40, approximate: false },
-    ]);
-
-    const paperPoints = inventory.sections.map((section) =>
-      section.parts.reduce((sum, part) => sum + (part.scoring.maxRawPoints ?? 0), 0)
-    );
-    expect(paperPoints).toEqual([100, 100, 100, 100]);
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([0.6, 0.6, 0.6, 0.6]);
-    expect(inventory.passRule).toMatchObject({ maximumPoints: 400, passPoints: 240 });
-  });
-
-  it("loads the project-defined Chinese pre-A1 HSK precursor with independent handwriting", () => {
-    const inventory = loadTaskShapeInventory("chinese", "pre-A1");
-    expect(inventory.target).toEqual({
-      name: "Coding Adventures Chinese pre-A1 Assessment — project-defined HSK precursor",
-      basis: "project-defined",
-    });
-    expect(inventory.sections.map((section) => section.skill)).toEqual([
-      "reading",
-      "listening",
-      "writing",
-      "speaking",
-    ]);
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([0.6, 0.6, 0.6, 0.6]);
-    expect(inventory.sections.find((section) => section.skill === "writing")?.parts.map((part) => part.id)).toEqual([
-      "prea1-writing-delayed-character-recall",
-      "prea1-writing-dictation-and-transcription",
-      "prea1-writing-bounded-production",
-    ]);
-    expect(inventory.passRule.note).toContain("not an official HSK level");
-  });
-
-  it("loads the official Spanish A1 performance target and its grouped pass rule", () => {
-    const inventory = loadTaskShapeInventory("spanish", "A1");
-    expect(inventory.target).toEqual({ name: "DELE A1", basis: "external" });
-    expect(inventory.sections.map((section) => section.skill)).toEqual([
-      "reading",
-      "listening",
-      "writing",
-      "speaking",
-    ]);
-    expect(inventory.sections.flatMap((section) => section.parts)).toHaveLength(13);
-    expect(inventory.administration).toMatchObject({
-      writtenMinutes: 95,
-      speakingMinutes: 10,
-      speakingPreparationMinutes: 10,
-    });
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([null, null, null, null]);
-
-    const reading = inventory.sections.find((section) => section.skill === "reading");
-    expect(reading?.parts.map((part) => part.items)).toEqual([5, 6, 6, 8]);
-    const listening = inventory.sections.find((section) => section.skill === "listening");
-    expect(listening?.parts.map((part) => part.replayCount)).toEqual([2, 2, 2, 2]);
-    const writing = inventory.sections.find((section) => section.skill === "writing");
-    expect(writing?.parts.map((part) => part.responseLength)).toEqual([
-      { unit: "words", minimum: 15, maximum: 25, approximate: false },
-      { unit: "words", minimum: 30, maximum: 40, approximate: false },
-    ]);
-  });
-
-  it("loads the external Arabic STAMP 4S target with a four-skill project A1 floor", () => {
-    const inventory = loadTaskShapeInventory("arabic", "A1");
-    expect(inventory.target).toEqual({
-      name: "Avant STAMP 4S Arabic (Modern Standard) — Level 3 / Novice-High project A1 floor",
-      basis: "external",
-    });
-    expect(inventory.sections.map((section) => section.skill)).toEqual([
-      "reading",
-      "listening",
-      "writing",
-      "speaking",
-    ]);
-    expect(inventory.sections.flatMap((section) => section.parts)).toHaveLength(4);
-    expect(inventory.administration).toMatchObject({
-      writtenMinutes: { minimum: 90, maximum: 105 },
-      speakingMinutes: { minimum: 20, maximum: 25 },
-      speakingPreparationMinutes: null,
-    });
-    expect(inventory.sections.map((section) => section.parts[0]?.items)).toEqual([30, 30, 3, 3]);
-    expect(inventory.sections.map((section) => section.parts[0]?.scoring.maxRawPoints)).toEqual([null, null, null, null]);
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([
-      1 / 3,
-      1 / 3,
-      3 / 8,
-      3 / 8,
-    ]);
-  });
-
-  it("loads the project-defined Latin A1 target with four independent thresholds", () => {
-    const inventory = loadTaskShapeInventory("latin", "A1");
-    expect(inventory.target).toEqual({
-      name: "Coding Adventures Latin A1 Assessment — project-defined equivalent",
-      basis: "project-defined",
-    });
-    expect(inventory.sections.map((section) => section.skill)).toEqual([
-      "reading",
-      "listening",
-      "writing",
-      "speaking",
-    ]);
-    expect(inventory.sections.flatMap((section) => section.parts)).toHaveLength(12);
-    expect(inventory.sections.map((section) =>
-      section.parts.reduce((sum, part) => sum + (part.scoring.maxRawPoints ?? 0), 0)
-    )).toEqual([25, 25, 25, 25]);
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([0.6, 0.6, 0.6, 0.6]);
-    expect(inventory.administration).toMatchObject({
-      writtenMinutes: 90,
-      speakingMinutes: 12,
-      speakingPreparationMinutes: 10,
-    });
-  });
-
-  it("loads the official French A1 performance target without flattening its forms", () => {
-    const inventory = loadTaskShapeInventory("french", "A1");
-    expect(inventory.target).toEqual({
-      name: "DELF A1 tout public",
-      basis: "external",
-    });
-    expect(inventory.sections.flatMap((section) => section.parts)).toHaveLength(14);
-    expect(inventory.administration.speakingMinutes).toEqual({ minimum: 5, maximum: 7 });
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([0.2, 0.2, 0.2, 0.2]);
-
-    const listening = inventory.sections.find((section) => section.skill === "listening");
-    expect(listening?.variants.map((variant) => variant.partIds.length)).toEqual([4, 5]);
-
-    const writing = inventory.sections.find((section) => section.skill === "writing");
-    expect(writing?.parts[1]?.responseLength).toEqual({
-      unit: "words",
-      minimum: 40,
-      maximum: null,
-      approximate: false,
-    });
-
-    const speaking = inventory.sections.find((section) => section.skill === "speaking");
-    expect(speaking?.parts.map((part) => part.scoring.maxRawPoints)).toEqual([4, 4, 4]);
-  });
-
-  it("loads DILF as French's sourced pre-A1 target without inventing per-skill floors", () => {
-    const inventory = loadTaskShapeInventory("french", "pre-A1");
-    expect(inventory.target).toEqual({ name: "DILF A1.1", basis: "external" });
-    expect(inventory.sections.map((section) => section.skill)).toEqual([
-      "reading",
-      "listening",
-      "writing",
-      "speaking",
-    ]);
-    expect(inventory.sections.map((section) => section.minutes)).toEqual([25, 25, 15, 10]);
-    expect(inventory.sections.flatMap((section) => section.parts)).toHaveLength(17);
-    expect(inventory.administration).toMatchObject({
-      writtenMinutes: 65,
-      speakingMinutes: 10,
-      speakingGroupMaximum: 1,
-      speakingPreparationMinutes: null,
-    });
-    expect(inventory.passRule).toMatchObject({
-      maximumPoints: 100,
-      passPoints: 50,
-      requiresEverySectionAttempted: false,
-    });
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([null, null, null, null]);
-    expect(inventory.sections.find((section) => section.skill === "speaking")?.parts.map((part) => part.id)).toEqual([
-      "speaking-price-transaction",
-      "speaking-present-person-or-place",
-      "speaking-express-need-or-request-information",
-      "speaking-describe-health-problem",
-    ]);
-  });
-
-  it("loads the official German A1 performance target", () => {
-    const inventory = loadTaskShapeInventory("german", "A1");
-    expect(inventory.target).toEqual({
-      name: "Goethe-Zertifikat A1: Start Deutsch 1",
-      basis: "external",
-    });
-    expect(inventory.sections.map((section) => section.skill)).toEqual([
-      "reading",
-      "listening",
-      "writing",
-      "speaking",
-    ]);
-    expect(inventory.sections.flatMap((section) => section.parts)).toHaveLength(11);
-    expect(inventory.sections.every((section) => section.variants.length === 0)).toBe(true);
-    expect(inventory.passRule).toMatchObject({ maximumPoints: 100, passPoints: 60 });
-    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([null, null, null, null]);
-  });
-
   it("keeps the corpus-wide research debt visible and round-robin", () => {
     const registry = loadLanguageRegistry();
     const present = listTaskShapeInventories();
     const backlog = buildTaskShapeBacklog(registry.languages.map((track) => track.id), present);
     const levels = ["pre-A1", "A1", "A2", "B1", "B2", "C1", "C2"];
     const registryOrder = registry.languages.map((track) => track.id);
-    expect(present).toEqual(expect.arrayContaining([
-      { language: "spanish", level: "A1" },
-      { language: "latin", level: "A1" },
-      { language: "french", level: "pre-A1" },
-      { language: "french", level: "A1" },
-      { language: "german", level: "A1" },
-      { language: "arabic", level: "A1" },
-      { language: "marwadi", level: "pre-A1" },
-      { language: "marwadi", level: "A1" },
-      { language: "gujarati", level: "pre-A1" },
-      { language: "chinese", level: "pre-A1" },
-    ]));
-    expect(new Set(present.map((item) => `${item.language}/${item.level}`)).size).toBe(present.length);
+    const presentIds = new Set(present.map((item) => `${item.language}/${item.level}`));
+    const backlogIds = new Set(backlog.map((item) => item.id.replace("task-shape/", "")));
+
+    expect(presentIds.size).toBe(present.length);
+    expect(backlogIds.size).toBe(backlog.length);
     expect(present).toEqual([...present].sort((left, right) => {
       const languageDifference = registryOrder.indexOf(left.language) - registryOrder.indexOf(right.language);
       return languageDifference || levels.indexOf(left.level) - levels.indexOf(right.level);
     }));
     expect(backlog).toHaveLength(registry.languages.length * levels.length - present.length);
+
+    for (const language of registryOrder) {
+      for (const level of levels) {
+        const id = `${language}/${level}`;
+        expect(Number(presentIds.has(id)) + Number(backlogIds.has(id))).toBe(1);
+      }
+    }
     for (const level of levels) {
       const presentAtLevel = present.filter((item) => item.level === level).length;
       expect(backlog.filter((item) => item.level === level)).toHaveLength(registry.languages.length - presentAtLevel);
     }
-    expect(backlog.some((item) => item.id === "task-shape/marwadi/pre-A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/marwadi/A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/gujarati/pre-A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/spanish/pre-A1")).toBe(true);
-    expect(backlog.some((item) => item.id === "task-shape/spanish/A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/spanish/A2")).toBe(true);
-    expect(backlog.some((item) => item.id === "task-shape/latin/A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/latin/A2")).toBe(true);
-    expect(backlog.some((item) => item.id === "task-shape/french/pre-A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/french/A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/french/A2")).toBe(true);
-    expect(backlog.some((item) => item.id === "task-shape/german/A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/german/A2")).toBe(true);
-    expect(backlog.some((item) => item.id === "task-shape/arabic/A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/arabic/A2")).toBe(true);
-    expect(backlog.some((item) => item.id === "task-shape/chinese/pre-A1")).toBe(false);
-    expect(backlog.some((item) => item.id === "task-shape/chinese/A1")).toBe(true);
   });
 
   it("rejects a missing skill instead of treating it as exam-ready", () => {
