+++ code/packages/typescript/human-language-data/tests/task-shapes/gujarati.test.ts
@@ -1,0 +1,33 @@
+import { describe, expect, it } from "vitest";
+import { loadTaskShapeInventory } from "../../src/loader.js";
+
+describe("Gujarati task-shape inventories", () => {
+  it("loads the project-defined Gujarati pre-A1 floor with four separate 100-point papers", () => {
+    const inventory = loadTaskShapeInventory("gujarati", "pre-A1");
+    expect(inventory.target).toEqual({
+      name: "Coding Adventures Gujarati pre-A1 Assessment — project-defined equivalent",
+      basis: "project-defined",
+    });
+    expect(inventory.sections.map((section) => section.skill)).toEqual([
+      "reading",
+      "listening",
+      "writing",
+      "speaking",
+    ]);
+    expect(inventory.administration).toMatchObject({
+      writtenMinutes: 32,
+      speakingMinutes: 8,
+      speakingPreparationMinutes: 0,
+    });
+    expect(inventory.sections.map((section) =>
+      section.parts.reduce((sum, part) => sum + (part.scoring.maxRawPoints ?? 0), 0)
+    )).toEqual([100, 100, 100, 100]);
+    expect(inventory.passRule).toMatchObject({ maximumPoints: 400, passPoints: 240 });
+    expect(Object.values(inventory.passRule.independentSkillThresholds)).toEqual([0.6, 0.6, 0.6, 0.6]);
+    expect(inventory.sections.find((section) => section.skill === "writing")?.parts.map((part) => part.id)).toEqual([
+      "prea1-writing-delayed-recall",
+      "prea1-writing-dictation",
+      "prea1-writing-bounded-production",
+    ]);
+  });
+});
